Organize Issue Services better (#180432)
The issue services were due for a clean up. They now have this shape:

* `platform/issue` has the `IssueMainService`, the service running on the Main process that Desktop takes advantage of
* `workbench/contrib/issue` has the contributions and the registered commands
* `workbench/services/issue` has the services that the renderer side can use to launch the issue reporter.

After this, I will work on getting out the raw IPC calls in the IssueMainService.
