+++ src/vs/workbench/contrib/issue/browser/issue.contribution.ts
@@ -10,7 +10,7 @@ import { LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle
 import { IProductService } from 'vs/platform/product/common/productService';
 import { Registry } from 'vs/platform/registry/common/platform';
 import { Extensions, IWorkbenchContributionsRegistry } from 'vs/workbench/common/contributions';
-import { WebIssueService } from 'vs/workbench/contrib/issue/browser/issueService';
+import { WebIssueService } from 'vs/workbench/services/issue/browser/issueService';
 import { IWorkbenchIssueService } from 'vs/workbench/services/issue/common/issue';
 import { BaseIssueContribution } from 'vs/workbench/contrib/issue/common/issue.contribution';
