+++ src/vs/workbench/services/issue/electron-sandbox/issueMainService.ts
@@ -3,11 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
-import { ICommonIssueService } from 'vs/platform/issue/common/issue';
+import { registerMainProcessRemoteService } from 'vs/platform/ipc/electron-sandbox/services';
+import { IIssueMainService } from 'vs/platform/issue/common/issue';
 
-export const IIssueService = createDecorator<IIssueService>('issueService');
-
-export interface IIssueService extends ICommonIssueService {
-	stopTracing(): Promise<void>;
-}
+registerMainProcessRemoteService(IIssueMainService, 'issue');
